refactor: remove redundant test modules and consolidate test cases for clarity
