+++ modules/actor-core/src/api/failure_telemetry.rs
@@ -14,20 +14,4 @@ pub(crate) use telemetry_builder_fn::TelemetryBuilderFn;
 pub use telemetry_context::TelemetryContext;
 
 #[cfg(test)]
-mod tests {
-  use crate::api::{
-    extensions::Extensions,
-    failure_telemetry::{FailureTelemetryBuilderShared, FailureTelemetryShared, TelemetryContext},
-    supervision::telemetry::NoopFailureTelemetry,
-  };
-
-  #[test]
-  fn telemetry_builder_shared_invokes_closure() {
-    let extensions = Extensions::new();
-    let builder = FailureTelemetryBuilderShared::new(|_ctx| FailureTelemetryShared::new(NoopFailureTelemetry));
-    let ctx = TelemetryContext::new(None, extensions);
-
-    let telemetry = builder.build(&ctx);
-    telemetry.with_ref(|_impl| {});
-  }
-}
+mod tests;
