+++ modules/actor-core/src/api/process/dead_letter.rs
@@ -89,28 +89,4 @@ impl<M> Default for DeadLetterHub<M> {
 }
 
 #[cfg(test)]
-mod tests {
-  use alloc::sync::Arc;
-  use core::sync::atomic::{AtomicBool, Ordering};
-
-  use super::*;
-  use crate::api::process::pid::SystemId;
-
-  #[test]
-  fn publishes_to_listeners() {
-    let pid = Pid::new(SystemId::new("sys"), crate::api::actor::ActorPath::new());
-    let mut hub = DeadLetterHub::new();
-    let flag = Arc::new(AtomicBool::new(false));
-    let flag_clone = Arc::clone(&flag);
-
-    let listener = ArcShared::new(move |_letter: &DeadLetter<&'static str>| {
-      flag_clone.store(true, Ordering::SeqCst);
-    })
-    .into_dyn(|f| f as &DeadLetterListener<_>);
-
-    hub.subscribe(listener);
-
-    hub.publish(&DeadLetter::new(pid, "msg", DeadLetterReason::UnregisteredPid));
-    assert!(flag.load(Ordering::SeqCst));
-  }
-}
+mod tests;
