+++ modules/actor-core/src/api/process/pid.rs
@@ -283,56 +283,4 @@ impl fmt::Display for Pid {
 }
 
 #[cfg(test)]
-mod tests {
-  use super::*;
-  use crate::api::actor::{ActorId, ActorPath};
-
-  #[test]
-  fn pid_display_formats_uri() {
-    let mut path = ActorPath::new();
-    path = path.push_child(ActorId(1));
-    path = path.push_child(ActorId(2));
-    let pid = Pid::new(SystemId::new("cellex"), path.clone())
-      .with_node(NodeId::new("node1", Some(2552)))
-      .with_tag(PidTag::new("v1"));
-
-    assert_eq!(pid.to_string(), "actor://cellex@node1:2552/1/2#v1");
-    assert_eq!(pid.path(), &path);
-  }
-
-  #[test]
-  fn parse_pid_from_uri() {
-    let pid = Pid::parse("actor://cellex@node1:2552/1/2#v1").expect("parse pid");
-    assert_eq!(pid.scheme(), "actor");
-    assert_eq!(pid.system().to_string(), "cellex");
-    assert_eq!(pid.node().unwrap().host(), "node1");
-    assert_eq!(pid.node().unwrap().port(), Some(2552));
-    assert_eq!(pid.tag().unwrap().to_string(), "v1");
-    let mut expected_path = ActorPath::new();
-    expected_path = expected_path.push_child(ActorId(1));
-    expected_path = expected_path.push_child(ActorId(2));
-    assert_eq!(pid.path(), &expected_path);
-  }
-
-  #[test]
-  fn parse_pid_without_node_or_tag() {
-    let pid = Pid::parse("actor://cellex/42").expect("parse pid");
-    assert!(pid.node().is_none());
-    assert!(pid.tag().is_none());
-    let mut path = ActorPath::new();
-    path = path.push_child(ActorId(42));
-    assert_eq!(pid.path(), &path);
-  }
-
-  #[test]
-  fn parse_pid_invalid_segment_fails() {
-    let err = Pid::parse("actor://cellex/foo").expect_err("should fail");
-    assert_eq!(err, PidParseError::InvalidPathSegment);
-  }
-
-  #[test]
-  fn parse_pid_invalid_port_fails() {
-    let err = Pid::parse("actor://cellex@node:notaport/1").expect_err("should fail");
-    assert_eq!(err, PidParseError::InvalidPort);
-  }
-}
+mod tests;
